--- src/actions/account.js
@@ -1,10 +1,0 @@
-import {
-    AUTH_PASSWORD_SET,
-} from '../constants/account';
-
-export const setPassword = (value) => {
-    return {
-        type: AUTH_PASSWORD_SET,
-        value,
-    };
-};
