+++ src/components/Footer/index.css
@@ -1,4 +1,3 @@
-
 .footer{
     height: 70px;
     background-color: white;
