+++ src/components/SideLogo/index.css
@@ -33,6 +33,7 @@
     padding-bottom: 5px;
     color: white;
     font-size: 12px;
+    letter-spacing: 5px;
 }
 
 .sideBarVersion {
