+++ src/components/TextField/index.js
@@ -9,7 +9,7 @@ const TextField = (props) => {
 
     return (
         <input
-            className={'text_field'}
+            className="text_field"
             id={props.id}
             name={props.name}
             placeholder={props.placeholder}
