--- src/constants/account.js
@@ -1,1 +1,0 @@
-export const AUTH_PASSWORD_SET = 'AUTH_PASSWORD_SET';
