+++ src/constants/application.js
@@ -1,3 +1,4 @@
 export const APPLICATION_LOADING_ERROR = 'APPLICATION_LOADING_ERROR';
 export const APPLICATION_LOADING_SUCCESS = 'APPLICATION_LOADING_SUCCESS';
 export const APPLICATION_LOADING_IN_PROGRESS = 'APPLICATION_LOADING_IN_PROGRESS';
+export const AUTH_PASSWORD_SET = 'AUTH_PASSWORD_SET';
