+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from 'redux';
 import application from './application';
-import account from './account';
 
 export default combineReducers({
     application,
-    account,
 });
